requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
passlib[bcrypt]
pyjwt
python-multipart
